Feat: Added two new classes to admin.py.
The classes added to admin.py is PostAdmin and CommentAdmin. This is a part of building the admin site, which makes it possible for the admin to list, search and filter out certain posts on the admin page.

It also makes it possible for the admin to display, filter and search for comments, as well as viewing if the comment is approved or not.