Feat: Updated views.py, style.ccs,
Added a new view to view.py, PostDetail. Makes it possible for the User to enter a blog post to read the entire post on a new page. Will also make it possible for the User to see if they have already liked/saved the post.

Updated style.css with some style for the bootstrap cards.