Added a new class to forms.py called UserBlogPost, this will make it possible for the user to add their own posts to the blog.
Added a new model class to models.py, called Trail.

Created new url paths for the
Post-like and post-save and add-post.

Created views for PostLike and PostSave, as well as AddPost.
The AddPost will make sure that the user is logged in to be able to add their own posts. The user will also be able to delete/edit posts.

The view for contact has also been created, next thing to create is contact.html.

Also created the views for the difficulties, the idea is that the user somehow should be able to sort the trails by difficulty preference.

Added the add_post.html and it’s content.