Feat: views.py, new model in the Post class, edit_post.html
Edited a lot in my AddPost view to because it didn’t work as it should. I couldn’t get it to submit, this was fixed by removing the form parameter from:
def getsuccessurl(self):
       return reverse('home')

The error message was also stating that I hadn’t made any migrations since I added the featured_image to the Post class. So I made a migration and then it worked.

Added a difficulty model to the Post class in models.py. The idea is to make it possible for the users to add a difficulty level to the posts they add. This will make it easier for people to see if they have the ability to visit a specifik trail.