Feat: models.py, forms.py, contact.html, contact_success.html,
Added ’city’ to my Post model, this is for the users to be able to add in their post what city the hike they add is in.
Also added the approved variable to the Post model, this will send the new user posts to the admin panel where the admin will have to approve the posts.

Added widgets to my ContactForm class and to the UserBlogPost form in forms.py.

Added a the html for the contactform, when the user has submitted their message in the contact form they will be redirected to a success page which is contact_success.html where they will get a confirmation that their message has been sent.