Feat: views.py, edit_post.html, urls.py, forms.py, index.html
Added a new view called EditPost, which will make it possible for the user to edit their own posts. Also created a template for editing user posts, which will only make it possible for the user to edit their OWN posts. If the post is not created by the user, the ”edit post” button will not be visible.

Added a url for edit_post.

Also added a new form to forms.py called EditForm, to keep the same structure as the other forms.

I have also been trying to fix my index.html. For some reason my bootstrap cards were not showing aligned horizontally, but instead they showed vertically on desktop and on mobile they showed vertically in different sizes and looked weird overall. Fixed this by replacing all the code related to the cards in this template. Now it looks like it should, it just needs som styling.