Feat: new delete view & small style changes
Added DeletePost to views.py, which will make it possible for the user to delete their OWN posts. The delete button is not going to be visible on the posts that is not written by the logged in user.
To this view I also added a deletepost template, which I copy pasted from add_post template and changed the essential details.
A new url was also created for this DeletePost view.

Except this I have also adjusted and changed some small styling details to post_detail.html, base.html and index.html.